requirements.txt
kaggle==1.5.6
matplotlib==2.2.2
networkx==2.0
nose==1.3.7
numpy==1.13.3
pandas==0.20.3
pytorch==0.4.1
scikit-learn==0.19.1
scipy==0.19.1
wget==3.2



